move login view
